Updated services and transactional classes

+++ intact-jami/src/main/java/uk/ac/ebi/intact/jami/dao/ComplexDao.java
@@ -61,7 +61,7 @@ public interface ComplexDao extends InteractorDao<IntactComplex>{
 
     public Collection<LifeCycleEvent> getLifeCycleEventsForComplex(String ac);
 
-    public long countParticipantsForComplex(String ac);
+    public int countParticipantsForComplex(String ac);
 
     public Collection<ModelledConfidence> getConfidencesForComplex(String ac);
