+++ intact-jami/src/main/java/uk/ac/ebi/intact/jami/dao/ParticipantDao.java
@@ -57,5 +57,5 @@ public interface ParticipantDao<F extends AbstractIntactParticipant> extends Int
 
     public Collection<Alias> getAliasesForParticipant(String ac);
 
-    public long countFeaturesForParticipant(String ac);
+    public int countFeaturesForParticipant(String ac);
 }
