+++ intact-jami/src/main/java/uk/ac/ebi/intact/jami/dao/PublicationDao.java
@@ -67,5 +67,5 @@ public interface PublicationDao extends IntactBaseDao<IntactPublication> {
 
     public Collection<LifeCycleEvent> getLifeCycleEventsForPublication(String ac);
 
-    public long countExperimentsForPublication(String ac);
+    public int countExperimentsForPublication(String ac);
 }
